+++ api/models/user-type.js
@@ -1,29 +1,29 @@
-const {DataTypes} = require('sequelize');
-const { sequelize } = require('./language');
+const { DataTypes } = require("sequelize");
+const { sequelize } = require("./language");
 
 const UserType = sequelize.define(
-    'usertype',
-    {
-        UserTypeID:{
-            type: DataTypes.INTEGER,
-            autoIncrement: true,
-            allowNull: false,
-            primaryKey: true,
-            unique: true
-        },
-        UserID:{
-            type: DataTypes.INTEGER,
-            allowNull: false,
-        },
-        TypeID:{
-            type: DataTypes.INTEGER,
-            allowNull:false
-        }
-    },
-    {
-        tableName:'usertype',
-        timestamps: false
-    }
+   "usertype",
+   {
+      UserTypeID: {
+         type: DataTypes.BIGINT,
+         autoIncrement: true,
+         allowNull: false,
+         primaryKey: true,
+         unique: true,
+      },
+      UserID: {
+         type: DataTypes.BIGINT,
+         allowNull: false,
+      },
+      TypeID: {
+         type: DataTypes.BIGINT,
+         allowNull: false,
+      },
+   },
+   {
+      tableName: "usertype",
+      timestamps: false,
+   }
 );
 
 module.exports = UserType;
